Redistribute `DApp` component to `MobileDApp` and `DesktopDapp`
