+++ src/shared/components/DApps/MobileDApp.tsx
@@ -1,0 +1,6 @@
+import React from "react"
+import MobileScreen from "ui/MobileScreen"
+
+export default function MobileDApp() {
+  return <MobileScreen />
+}
